pyproject.toml
[project]
name = "urosim"
version = "0.1.0"
description = "UroSim — Python side of the ureteroscopy simulator (Unity handles physics + rendering)."
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
    "numpy",
    "scipy",
    "trimesh",
    "networkx",
    "scikit-image",
    "pymeshlab",
    "grpcio",
    "grpcio-tools",
    "protobuf",
    "Pillow",
]

[project.optional-dependencies]
core = ["numpy", "scipy", "trimesh", "networkx", "scikit-image", "pymeshlab"]
client = ["urosim[core]", "grpcio", "protobuf", "Pillow"]
eval = ["urosim[core]", "matplotlib"]
agents = ["urosim[client]", "openai"]
sim2real = ["diffusers", "transformers", "accelerate", "safetensors", "controlnet-aux"]
all = ["urosim[agents]", "urosim[eval]", "urosim[sim2real]"]

[dependency-groups]
dev = ["pytest", "mypy", "ruff"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["urosim"]

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = ["urosim/client/generated"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
# Generated protobuf/gRPC stubs use dynamic attribute loading that mypy
# cannot see statically. Exclude them from the source set and treat
# imports from that package as Any.
exclude = ["urosim/client/generated/"]

[[tool.mypy.overrides]]
module = ["urosim.client.generated.*"]
follow_imports = "skip"
ignore_errors = true
